query k8s metrics API to check capacity before invoking and misc. fixes

+++ resource-manager/app/src/main/java/com/tomcoward/heterogeneousfaas/resourcemanager/database/tables/FunctionExecutionTable.java
@@ -30,8 +30,8 @@ public class FunctionExecutionTable implements IDBTable {
                     .withColumn("function_name", DataTypes.ASCII)
                     .withColumn("worker", DataTypes.ASCII)
                     .withColumn("input_size", DataTypes.INT)
-                    .withColumn("duration", DataTypes.FLOAT)
-                    .withColumn("predicted_duration", DataTypes.FLOAT)
+                    .withColumn("duration", DataTypes.DOUBLE)
+                    .withColumn("predicted_duration", DataTypes.DOUBLE)
                     .withColumn("is_success", DataTypes.BOOLEAN)
                     .build();
